Move graph metrics to a sep dir
