+++ examples/synthetic_graph_evolution/abstract_graph_search.py
@@ -4,7 +4,7 @@ from io import StringIO
 from itertools import product
 from typing import Sequence, Type
 
-from examples.synthetic_graph_evolution.graph_metrics import *
+from golem.metrics.graph_metrics import *
 from examples.synthetic_graph_evolution.utils import draw_graphs_subplots
 from golem.core.adapter.nx_adapter import BaseNetworkxAdapter, nx_to_directed
 from golem.core.dag.verification_rules import has_no_self_cycled_nodes
