+++ examples/synthetic_graph_evolution/utils.py
@@ -6,7 +6,7 @@ import networkx as nx
 import numpy as np
 from networkx import gnp_random_graph
 
-from examples.synthetic_graph_evolution.graph_metrics import spectral_dists_all
+from golem.metrics.graph_metrics import spectral_dists_all
 from golem.core.adapter.nx_adapter import BaseNetworkxAdapter
 from golem.core.optimisers.opt_history_objects.opt_history import OptHistory
 from golem.visualisation.graph_viz import GraphVisualizer
